made build
build description not available

+++ maventry/src/proj/server/servicepack/EmployeeServiceImpl.java
@@ -22,37 +22,78 @@ public class EmployeeServiceImpl implements EmployeeService {
 	}
 	
 	@Override
-	public String addEmployee(Employee toBeAdded){	
+	public String addEmployee(Employee toBeAdded) {
+		
+		if(checkPermission(MethodPermissions.WRITE_EMPLOYEE.name())) {
+			return "RoleNotAllowed";
+		}
+		
+		
 		return empdao.addEmployee(toBeAdded);
-	};
+	}
 
 	@Override
 	public List<Employee> listEmployees() {
 		
+		if(checkPermission(MethodPermissions.READ_EMPLOYEE.name())) {
+			return null;
+		}
+		
+		
 		return empdao.listEmployees();
-	};
+	}
    
     @Override
-	public void updateEmployee(int id, String name, int salary) {
+	public String updateEmployee(int id, String name, int salary,String designation,String password) {
     	
-    	empdao.updateEmployee(id, name, salary);
-    };
+    	if(checkPermission(MethodPermissions.UPDATE_EMPLOYEE.name())) {
+			return "not updated due to not permissions";
+		}
+    	
+    	empdao.updateEmployee(id, name, salary,designation,password);
+    	return "Updated Employee";
+    }
     
     @Override
 	public boolean deleteEmployee(String remEmp) {
-
+    	
+    	if(checkPermission(MethodPermissions.WRITE_EMPLOYEE.name())) {
+			return false;
+		}
+    	
     	
     	return empdao.deleteEmployee(remEmp);
-    };
+    }
  
     @Override
 	public Employee getEmployee(String username){
+    	
+    	if(checkPermission(MethodPermissions.READ_EMPLOYEE.name())) {
+			return null;
+		}
+    	
     	return empdao.getEmployee(username);
-    };
+    }
    
     @Override
 	public List<Employee> getEmployeesOnDesig(String designame){
+    	
+    	if(checkPermission(MethodPermissions.READ_EMPLOYEE.name())) {
+			return null;
+		}
+    	
+    	
     	return empdao.getEmployeesOnDesig(designame);
     };
+    
+    public Boolean checkPermission(String authority) {
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		for(GrantedAuthority ga: auth.getAuthorities()) {
+			if(ga.getAuthority().equals(authority)) {
+				return false;
+			}
+		}
+		return true;
+	}
 	
 }
